extract vertex duplication in MRVertDuplication.h/.cpp
