+++ source/MRMesh/MRMeshBuilder.h
@@ -3,7 +3,7 @@
 #include "MRMeshBuilderTypes.h"
 #include "MRMesh.h"
 #include "MRProgressCallback.h"
-#include <utility>
+#include "MRVertDuplication.h"
 
 namespace MR
 {
@@ -36,18 +36,6 @@ namespace MeshBuilder
 /// construct mesh topology from a set of triangles with given ids;
 /// if skippedTris is given then it receives all input triangles not added in the resulting topology
 MRMESH_API MeshTopology fromTriangles( const Triangulation & t, const BuildSettings & settings = {}, ProgressCallback progressCb = {} );
-
-struct VertDuplication
-{
-    VertId srcVert; ///< original vertex before duplication
-    VertId dupVert; ///< new vertex after duplication
-};
-
-/// resolve non-manifold vertices by creating duplicate vertices in the triangulation (which is modified)
-/// `lastValidVert` is needed if `region` or `t` does not contain full mesh, then first duplicated vertex will have `lastValidVert+1` index
-/// return number of duplicated vertices
-MRMESH_API size_t duplicateNonManifoldVertices( Triangulation & t, FaceBitSet * region = nullptr,
-    std::vector<VertDuplication>* dups = nullptr, VertId lastValidVert = {} );
 
 /// construct mesh topology from a set of triangles with given ids;
 /// unlike simple fromTriangles() it tries to resolve non-manifold vertices by creating duplicate vertices;
@@ -124,31 +112,6 @@ MRMESH_API int uniteCloseVertices( Mesh & mesh, float closeDist, bool uniteOnlyB
 /// then the mesh is rebuilt from the remaining triangles
 /// \return the number of vertices united, 0 means no change in the mesh
 MRMESH_API int uniteCloseVertices( Mesh& mesh, const UniteCloseParams& params = {} );
-
-/// classification of the triangles around one vertex
-struct VertInfo
-{
-    /// the number of chains of connected triangles around the vertex;
-    /// numChains is set to 0 if numRepeatedVerts > 0
-    int numChains = 0;
-
-    /// the number of vertices, which are passed more than once
-    int numRepeatedVerts = 0;
-};
-
-/// describes a vertex and one of triangles incident to it
-struct VertTri
-{
-    VertId v;
-    FaceId f;
-
-    auto asPair() const { return std::make_pair( v, f ); }
-    friend bool operator <( const VertTri& l, const VertTri& r ) { return l.asPair() < r.asPair(); }
-};
-
-/// computes VertInfo of one vertex given all its incident triangles in [begin, end), all referencing the same vertex;
-/// allocates temporary hash maps on every call, so prefer batch processing over calling it per vertex of a large mesh
-[[nodiscard]] MRMESH_API VertInfo inspectVertNeighbourhood( const Triangulation & t, const VertTri * begin, const VertTri * end );
 
 } //namespace MeshBuilder
